Add TextOutBox

+++ src/main_scene.modules/model/TileHoveringModule.cpp
@@ -10,9 +10,9 @@ namespace Narradia {
         auto player_pos = Player::Get()->position();
         auto col = mouse_pos.x / tile_width;
         auto row = mouse_pos.y / tile_height;
-        auto map_x = player_pos.x - 5 + col;
-        auto map_y = player_pos.y - 5 + row;
-        
+        auto map_x = static_cast<int>(player_pos.x - 5 + col);
+        auto map_y = static_cast<int>(player_pos.y - 5 + row);
+
         hovered_tile_ = {map_x, map_y};
     }
 }
